+++ projects/mind-mapp/app/src/utils/index.ts
@@ -7,6 +7,7 @@ export * from './fitViewMath';
 export * from './id';
 export * from './importValidation';
 export * from './sampleMap';
+export * from './searchNodes';
 export * from './selectionBounds';
 export * from './selectionText';
 export * from './shortcuts';